libmkvparser: use cues even when not all clusters loaded
Previously, if the client requested a seek time greater than what was already
(pre)loaded in the cluster cache, then we would give up immediately and revert
to searching over the cluster index.

We're now smarter about searching over the cue points.  If the client
specifieds a time outside what we have in cache, then we search for
the cue point having a timecode not greater than the last cluster
in the cache.

Change-Id: I3b45115015ed5ec99961e3efd07386fd42e63119