libmkvparser: parse cues too, if they're at the front
The source filter will parse the cues, no matter where they are in the file,
as long as the cues element has an entry in the seekhead element.

The splitter filter doesn't work like that.  It doesn't bother with the
seekhead, because the seekhead entries can designate positions in the
file that haven't been loaded yet (and hence are not accessible).

What it does do is parse the header elements that the front of the file,
ahead of the first cluster.  If the cues element is there (it can go
either at the front before the clusters, or at the back after them),
then it parses the cues element too.

Change-Id: I31d1297f2a85fdbd28ae52a595bd3c73e2740749