load up to next cue point for large seek
The parser loads clusters incrementally, building up an index.
Seek requests always work over what's in the cache.

If the seek request names a time that corresponds to a cluster
that is already loaded in the cache, then all is well and we
seek to that cluster.

If the seek request names a time that is beyond what's in
the cache, then previously the seek would find the largest
cue point that was actually in the cache, and seek to that.

This new revision finds the cue point (if any) that lies
immediately beyond the last loaded cluster, and then preloads
clusters until that cue point has been loaded.

Note that this does not load every cluster up to the seek
point, as they could be potentially expensive, if the
seek request is far beyond what's loaded in the cache.
But this is a comprimise: we at least load something new,
so have something different to display to the user.

Change-Id: I4dc92c858f742a1995a5d749946028b6602123d6