Add support for altref to VP8 encoder filter.
Added the interface to turn altref frames on and off in the VP8 encoder
filter. Added the interface to change the values used on the altref
filter. I also changed the timestamps to be scaled down to milliseconds
before they are sent into the vp8 encoder. This is to insure the WebM
muxer will have strictly increasing timestamps with altref frames. Also
updated libvpx to v0.9.2-114-gbdf469c to get the newest code for altref
frames.

Change-Id: I4ba99439e266c0b73593b21e16a9abab704e657a