webmmfsource: fixed bug in loading next cluster
After we parse the next cluster, we load the next cluster in cache;
this allows us to safely load the cluster's block entries.

To populate the cache, we specify a position and length.  However,
we were specifying the position of the start of the cluster, and
the length of its payload -- but this does not take into account
the length of the EBML ID and Size fields.

We now request 16 extra bytes when loading the next cluster, which
ensures that the entire cluster (ID + Size + Payload) has been
loaded into the cache.

Change-Id: I5c9591f83c42bd653b1284bb48ecb9810b7df61b