webmmf: removed dependency in omahautil
The implementation file for omahautil had an obsolete dependency
on memutil; that has been removed.

The header file included comdef.h but that's not really the
file you want (it contains the support for COM exceptions and smart
pointers).  The objbase.h file was included instead (it contains
the declarations for HRESULT and GUID, etc).

Change-Id: I8d7216575b2fd4ff668a2731e09a7434e4e12575