vp8decoder: tolerate vp8 decode errors
Previously, when the vp8 library call to decode the frame failed, the
filter would return S_FALSE and essentially stop the video stream.

In order to be more robust in the presence of errors in the stream,
the filter now transitions to an intermediate state when a decode
failure has occured.  It waits until a keyframe has been successfully
decoded, and then transitions back to its normal state.

Change-Id: I8db0149f132c80812849466bca96432844b989f8