WebmMfVorbisDec: incorrect assertion
The pipeline can ask the filter about a change in output format.  The
filter returns an answer about whether the format change would be
acceptable.

If the request is "invalid" this not really an error.  The pipeline is
just asking about a potential output format, and so the filter answers
no.  The presence of the assert is incorrect, since if there's a
difference in sampling rates (the case here), then the filter merely
returns an answer that such a change would not be allowed.

Change-Id: I868f738d332cbf3e9486035e9675ef9b6918a3e3