mkvparserstream: relax constraint on base time
We use a "base" time for seeking.  When we initialize, the
base time is set to the time of the first block in the
cluster.

There existed some checks to assert that the base time was
non-negative.  However, this caused problems when the very
first frame of the file had a negative timestamp.  (This is
not a legal Matroska file, but we have to at least handle
this case by doing someting sensible.)

We do two things to handle the issue of negative timestamps.
First we throw away any frames with a negative timestamp,
continuing the parse until we find a non-negative frame.
Secondly, we relax the constraint that the base time must be
non-negative.

Change-Id: I5c9dfc8be6e0ad5215c42ce595279749a3d86149