VP8Encoder: added VP9 support
Added new interface to VP8 encoder (IVPXEncoder), that allows client
to specify whether this encode is VP8 (the default) or VP9.

The webm muxer filter was also modified to accept VP9 as the media
subtype.

A new command-line switch ("-encoder") was added to specify "vp8" or
"vp9" as the type for the encode.

Change-Id: I96f04fd2aa0622ca94e478e6648cfe7d44a7f9dc